HSLdevcom/bultti#98 New editEquipmentCatalogue component, enable adding additional equipment catalogues.
