+++ src/report/ReportTableFilters.tsx
@@ -6,7 +6,6 @@ import { omit } from 'lodash'
 import { ControlGroup } from '../common/components/form'
 import Dropdown from '../common/input/Dropdown'
 import { Button, ButtonSize, ButtonStyle, RemoveButton } from '../common/components/Button'
-import { CrossThick } from '../common/icon/CrossThick'
 import { text } from '../util/translate'
 import { FlexRow } from '../common/components/common'
 import { SubHeading } from '../common/components/Typography'
